Cleaned up, separated and documented the Extension methods. Still needs a second pass for individual methods / properties.
